+++ ssr/feature/api-v1/routes/v1-document-create.js
@@ -1,3 +1,6 @@
+const {
+  computeNextIteration,
+} = require('fetchq/lib/utils/compute-next-iteration');
 const schema = require('./v1-document-create.schema');
 
 const v1DocumentCreate = {
@@ -12,9 +15,10 @@ const v1DocumentCreate = {
       const { queue } = params;
       const { subject } = body;
       const payload = JSON.stringify(body.payload).replace(/'/g, "''''");
+      const next_iteration = computeNextIteration(body.next_iteration);
 
       const _sqlPush = subject
-        ? `SELECT * FROM "fetchq"."doc_push"('${queue}', '${subject}', 0, 0, NOW(), '${payload}');`
+        ? `SELECT * FROM "fetchq"."doc_push"('${queue}', '${subject}', 0, 0, ${next_iteration}, '${payload}');`
         : `SELECT * FROM "fetchq"."doc_append"('${queue}', '${payload}');`;
       const resPush = await fetchq.pool.query(_sqlPush);